Closes #104 - Adding wish counter ui

+++ src/app/shared/components/book/book.component.ts
@@ -57,6 +57,7 @@ export class BookComponent implements OnInit {
   disabledButton = false;
   previousBooksPage: booksPage;
   rating = 0;
+  public clickCounter = 0;
 
   constructor(
     private translate: TranslateService,
@@ -472,30 +473,39 @@ export class BookComponent implements OnInit {
   }
 
   public changeWishList(book: IBook): void {
-    if (this.isWished) {
-      this.wishListService.removeFromWishList(book.id).subscribe(
-        (data) => {
-          this.isWished = false;
-        },
-        (error) => {
-          this.notificationService.error(
-            this.translate.instant('Something went wrong'),
-            'X'
-          );
-        }
-      );
-    } else {
-      this.wishListService.addToWishList(book.id).subscribe(
-        (data) => {
-          this.isWished = true;
-        },
-        (error) => {
-          this.notificationService.error(
-            this.translate.instant('Cannot add own book to the wish list'),
-            'X'
-          );
-        }
-      );
+    this.clickCounter += 1;
+    if (this.clickCounter <= 1) {
+      if (this.isWished) {
+        book.wishCount -= 1;
+        this.wishListService.removeFromWishList(book.id).subscribe(
+          (data) => {
+            this.isWished = false;
+            this.clickCounter = 0;
+          },
+          (error) => {
+            this.clickCounter = 0;
+            this.notificationService.error(
+              this.translate.instant('Something went wrong'),
+              'X'
+            );
+          }
+        );
+      } else {
+        book.wishCount += 1;
+        this.wishListService.addToWishList(book.id).subscribe(
+          (data) => {
+            this.isWished = true;
+            this.clickCounter = 0;
+          },
+          (error) => {
+            this.clickCounter = 0;
+            this.notificationService.error(
+              this.translate.instant('Cannot add own book to the wish list'),
+              'X'
+            );
+          }
+        );
+      }
     }
   }
